LayoutTests/crypto/subtle/ecdh-import-raw-key-p256-expected.txt

Test importing an ECDH raw key with P-256

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
PASS publicKey.toString() is '[object CryptoKey]'
PASS publicKey.type is 'public'
PASS publicKey.extractable is true
PASS publicKey.algorithm.name is 'ECDH'
PASS publicKey.algorithm.namedCurve is 'P-256'
PASS publicKey.usages is [ ]
PASS successfullyParsed is true

TEST COMPLETE
